# doc/swagger.yaml
definitions:
...
  exchange.ClaimFormat:
    properties:
      jwt:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vc:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vp:
        $ref: '#/definitions/exchange.JWTType'
      ldp:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vc:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vp:
        $ref: '#/definitions/exchange.LDPType'
    type: object
...
  exchange.JWTType:
    properties:
      alg:
        items:
          type: string
        type: array
    required:
    - alg
    type: object
  exchange.LDPType:
    properties:
      proof_type:
        items:
          type: string
        type: array
    required:
    - proof_type
    type: object
  exchange.PresentationDefinition:
    properties:
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      frame:
        description: https://identity.foundation/presentation-exchange/#json-ld-framing-feature
      id:
        type: string
      input_descriptors:
        items:
          $ref: '#/definitions/exchange.InputDescriptor'
        type: array
      name:
        type: string
      purpose:
        type: string
      submission_requirements:
        items:
          $ref: '#/definitions/exchange.SubmissionRequirement'
        type: array
    required:
    - id
    - input_descriptors
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestRequest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      issuerDid:
        type: string
      issuerName:
        type: string
      name:
        type: string
      outputDescriptors:
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentationDefinition:
        $ref: '#/definitions/exchange.PresentationDefinition'
    required:
    - format
    - issuerDid
    - outputDescriptors
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
      manifestJwt:
        type: string
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetApplicationResponse:
    properties:
      application:
        $ref: '#/definitions/manifest.CredentialApplication'
      id:
        type: string
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.GetApplicationsResponse:
    properties:
      applications:
        items:
          $ref: '#/definitions/manifest.CredentialApplication'
        type: array
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
      id:
        type: string
      manifestJwt:
        type: string
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestsResponse:
    properties:
      manifests:
        items:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestResponse'
        type: array
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetResponseResponse:
    properties:
      id:
        type: string
      response:
        $ref: '#/definitions/manifest.CredentialResponse'
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.GetResponsesResponse:
    properties:
      responses:
        items:
          $ref: '#/definitions/manifest.CredentialResponse'
        type: array
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.Operation:
    properties:
      done:
        description: Whether this operation has finished.
        type: boolean
      id:
        description: The name of the resource related to this operation. E.g. "presentations/submissions/<uuid>"
        type: string
      result:
        $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.OperationResult'
        description: Populated if Done == true.
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.OperationResult:
    properties:
      error:
        description: Populated when there was an error with the operation.
        type: string
      response:
        description: Populated iff Error == "". The type should be specified in the
          calling APIs documentation.
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.SubmitApplicationRequest:
    properties:
      applicationJwt:
        type: string
    required:
    - applicationJwt
    type: object
...
  manifest.CredentialApplication:
    properties:
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      id:
        type: string
      manifest_id:
        type: string
      presentation_submission:
        $ref: '#/definitions/exchange.PresentationSubmission'
        description: Must be present if the corresponding manifest contains a presentation_definition
      spec_version:
        type: string
    required:
    - format
    - id
    - manifest_id
    - spec_version
    type: object
  manifest.CredentialManifest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      id:
        type: string
      issuer:
        $ref: '#/definitions/manifest.Issuer'
      name:
        type: string
      output_descriptors:
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentation_definition:
        $ref: '#/definitions/exchange.PresentationDefinition'
      spec_version:
        type: string
    required:
    - id
    - issuer
    - output_descriptors
    - spec_version
    type: object
  manifest.CredentialResponse:
    properties:
      application_id:
        type: string
      denial:
        properties:
          input_descriptors:
            items:
              type: string
            type: array
          reason:
            type: string
        required:
        - reason
        type: object
      fulfillment:
        properties:
          descriptor_map:
            items:
              $ref: '#/definitions/exchange.SubmissionDescriptor'
            type: array
        required:
        - descriptor_map
        type: object
      id:
        type: string
      manifest_id:
        type: string
      spec_version:
        type: string
    required:
    - id
    - manifest_id
    - spec_version
    type: object
  manifest.Issuer:
    properties:
      id:
        type: string
      name:
        type: string
      styles:
        $ref: '#/definitions/rendering.EntityStyleDescriptor'
        description: |-
          an object or URI as defined by the DIF Entity Styles specification
          https://identity.foundation/wallet-rendering/#entity-styles
    required:
    - id
    type: object
  manifest.OutputDescriptor:
    properties:
      description:
        type: string
      display:
        $ref: '#/definitions/rendering.DataDisplay'
        description: 'both below: an object or URI as defined by the DIF Entity Styles
          specification'
      id:
        description: Must be unique within a manifest
        type: string
      name:
        type: string
      schema:
        type: string
      styles:
        $ref: '#/definitions/rendering.EntityStyleDescriptor'
    required:
    - id
    - schema
    type: object
...
  rendering.DataDisplay:
    properties:
      description:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      properties:
        items:
          $ref: '#/definitions/rendering.LabeledDisplayMappingObject'
        type: array
      subtitle:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      title:
        $ref: '#/definitions/rendering.DisplayMappingObject'
    type: object
...
  rendering.EntityStyleDescriptor:
    properties:
      background:
        $ref: '#/definitions/rendering.ColorResource'
      hero:
        $ref: '#/definitions/rendering.ImageResource'
      text:
        $ref: '#/definitions/rendering.ColorResource'
      thumbnail:
        $ref: '#/definitions/rendering.ImageResource'
    type: object
...
host: localhost:8080
info:
  contact:
    email: tbd-developer@squareup.com
    name: TBD
    url: https://github.com/TBD54566975/ssi-service/issues
  description: https://github.com/TBD54566975/ssi-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: "0.1"
paths:
...
  /v1/manifests:
    get:
      consumes:
      - application/json
      description: Checks for the presence of a query parameter and calls the associated
        filtered get method
      parameters:
      - description: string issuer
        in: query
        name: issuer
        type: string
      - description: string schema
        in: query
        name: schema
        type: string
      - description: string subject
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get manifests
      tags:
      - ManifestAPI
    put:
      consumes:
      - application/json
      description: Create manifest
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create manifest
      tags:
      - ManifestAPI
  /v1/manifests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete manifest by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete manifests
      tags:
      - ManifestAPI
    get:
      consumes:
      - application/json
      description: Get a credential manifest by its id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get manifest
      tags:
      - ManifestAPI
  /v1/manifests/applications:
    get:
      consumes:
      - application/json
      description: Checks for the presence of a query parameter and calls the associated
        filtered get method
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetApplicationsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get applications
      tags:
      - ApplicationAPI
    put:
      consumes:
      - application/json
      description: Submit a credential application in response to a credential manifest.
        The request body is expected to
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.SubmitApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Operation with a SubmitApplicationResponse type in the `result.response`
            field.
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.Operation'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Submit application
      tags:
      - ApplicationAPI
  /v1/manifests/applications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete application by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete applications
      tags:
      - ApplicationAPI
    get:
      consumes:
      - application/json
      description: Get application by id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetApplicationResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get application
      tags:
      - ApplicationAPI
  /v1/manifests/responses:
    get:
      consumes:
      - application/json
      description: Checks for the presence of a query parameter and calls the associated
        filtered get method
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetResponsesResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get responses
      tags:
      - ResponseAPI
  /v1/manifests/responses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete response by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete responses
      tags:
      - ResponseAPI
    get:
      consumes:
      - application/json
      description: Get response by id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetResponseResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get response
      tags:
      - ResponseAPI
...
swagger: "2.0"